Bug 1669181 - Part 45: Split ScriptFromCalleeToken from "jit/JitFrames.h". r=jandem
Split `ScriptFromCalleeToken` from "jit/JitFrames.h" to avoid the dependency
on "vm/JSFunction.h".

Differential Revision: https://phabricator.services.mozilla.com/D92738
