+++ js/src/jit/JitScript.cpp
@@ -15,6 +15,7 @@
 #include "jit/BaselineIC.h"
 #include "jit/BytecodeAnalysis.h"
 #include "jit/IonScript.h"
+#include "jit/ScriptFromCalleeToken.h"
 #include "util/Memory.h"
 #include "vm/BytecodeIterator.h"
 #include "vm/BytecodeLocation.h"
