+++ js/src/jit/arm/Bailouts-arm.cpp
@@ -8,6 +8,7 @@
 #include "jit/Bailouts.h"
 #include "jit/JitRuntime.h"
 #include "jit/SafepointIndex.h"
+#include "jit/ScriptFromCalleeToken.h"
 #include "vm/JSContext.h"
 #include "vm/Realm.h"
