+++ js/src/jit/mips32/Bailouts-mips32.cpp
@@ -7,6 +7,7 @@
 #include "jit/mips32/Bailouts-mips32.h"
 
 #include "jit/JitRuntime.h"
+#include "jit/ScriptFromCalleeToken.h"
 #include "vm/JSContext.h"
 #include "vm/Realm.h"
