+++ js/src/jit/x86/Bailouts-x86.cpp
@@ -7,6 +7,7 @@
 #include "jit/Bailouts.h"
 #include "jit/JitRuntime.h"
 #include "jit/SafepointIndex.h"
+#include "jit/ScriptFromCalleeToken.h"
 #include "vm/JSContext.h"
 #include "vm/Realm.h"
